requirements.txt
streamlit
pandas
requests
beautifulsoup4
lxml
python-dateutil
transformers
torch
altair